Fix a problem where pagination of results could miss CPO collection
If a user had hundreds of files, the naive search for the CPO collection among
all filenames could miss, and cause a new, empty directory to be created.

Now, as long as a user has fewer than <page-limit> code.pyret.org directories,
things will work.  <page-limit> appears to be around 100