Restore credentials immediately after sending request
This was a (obvious in hindsight) bug with async timing; if we wait until the
request returns to restore the credential, we have races with other
credential-needing requests that are sent while that request is out.

If we restore the credential in the same turn as the request is sent, this
works better.  Fixes issues with transitive my-gdrive/shared-gdrive includes
clobbering one another's credential state.