Move notification area into #main rather than #header
I'm fairly sure that since #header has (intentionally) "overflow: hidden"
set, and #notifications was inside #header DOM-wise, that Safari (and
sometimes Chrome) didn't know what to do with the combination of the overflow
rule and "position: absolute" placing the element outside the x/y box of the
header.  Now it's inside #main, which doesn't have overflow: hidden set and
ought to allow rendering anywhere.  notificationArea now needs a high z-index
to be shown on top of interactions, as well.

Also renames a variable from err->msg.

Closes #561.