Add new /ide endpoint with html-webpack-plugin integration
This change creates a new /ide endpoint where we will be integrating all
the new UI components. The html for this file is generated using the
webpack-html-plugin, which automatically adds the correct <script> tags
to the generated html.

There is also now configuration for the webpack-dev-server which will
automatically recompile any changed javascript and serve it up on port
8080.

To get to the new ide, you can run:

```
$ heroku local:run `npm bin`/webpack-dev-server
```

And go to http://localhost:8080/views/ide.html

Or you can get to it through the regular pyret server at

http://localhost:5000/ide