calculate the split between repl and text editor
now we can tell which half of the display we are looking at.
One thing to check out is how would we tell which line number
the person is looking at?
Would we have to know the height of one line of text
and the height of the top bar,
then calculate it from there
using the first line of the text editor as the base offset?
How would we do this with error messages, though?