move setgaze listener stuff to repl-ui
there is the issue that webgazer is not paused at the beginning.
I think we can only pause when we are collecting data,
and collection does not start until the user has clicked.
So if we pause between webgazer.begin and the user clicking,
this pause seems to be lost.