error message rendering rework Drastic simplification of error message rendering.
Unstyled marks do not require a codemirror display update, so cmlocs are
replaced with a `Position` class that keeps track of its location with
an unstyled TextMarker. Highlights are added and removed by adding and
removing TextMarkers.

This negates the need for complex batching of highlights, and eliminates
all bugs caused by using a stale position for a range, rather than an
up-to-date position.

Highlights can be displayed eagerly or lazily, with multiple colors or
just a single color. Preferences are chached in sessionStorage, and
highlighting mode can be switched without re-running error messages.

this fixes brownplt/pyret-lang#803 and other untracked (at the time of
writing this message) bugs.