Remove GOOGLE_SERVER_KEY, in favor of using alt=media and a client_key
It looks like the gapis, at some point in the last year, got the ability to
request the content of a public file from the client.  Good job Drive!

For fetching shared files, now use the "preferred" alt=media mechanism here:

https://developers.google.com/drive/v2/web/manage-downloads

This is a win all around, as it allows us to skip a proxying round-trip to the
server and produce the text of the file faster.

This also avoids having the downloadGoogleFile endpoint, which relied on a
to-be-deprecated mechanism for fetching file content:

http://googleappsdeveloper.blogspot.com/2015/08/deprecating-web-hosting-support-in.html

This does not fix the issue that students in some GAE settings can't create
public-on-the-web files in the first place, but that can be separately
addressed.