Change for caching locators per 602642b on pyret-lang
Implement the caching behavior described in 602642b on pyret-lang, so that the
extent of the locator cache is the same as the extent of a definitions +
interactions run.

Also make the URI for my-gdrive imports just have the filename.  Since all
my-gdrive imports live in the same directory necessarily, the ID doesn't add
any uniqueness.  This also makes it so the URI for a module can be determined
without doing a costly fetch to Google Drive to match the name to the ID.