Loosening the restrictions on color components
Since it would be very annoying to have to clamp or truncate r/g/b values to integers, and defining algebra over colors really wants the components to Just Work Normally, we're loosening the type restrictions on color components to be arbitrary numbers.  When *using* the color for images, r/g/b will be converted to fixnums, clamped to the range [0,255], and then floored to an int.  Alpha will be converted to a fixnum and clamped to the range [0.0, 1.0].  Rendering the color at the REPL will draw a swatch by using the color as above, and then will display the numbers as they are actually contained in the color value, for the other renditions of the color.