Make published files not show in dashboard
The fix has a few parts, with future-proofing in mind.

1. Ideally, I would like to just use the 'originalProgram' key to filter out
any files that are shared. Unfortunately, I can't find a way to write that
query. This issue is worth tracking:

https://stackoverflow.com/questions/23900988/how-to-search-google-drive-for-file-with-a-specific-property/26286007

As a result, we need some other way to detect shared files in this query.

2. Right now, every user who isn't circumventing defaults has all shared files
in code.pyret.org.shared. So, we can filter out files that are in that
directory as part of the query, which is easy enough.

3. However, observe that if, at the beginning of CPO, we'd also added a flag
property with a known value, we could have just queried based on that. I added
this property to the server, so that in the future we can actually write these
queries. I think that this is important, because we'd eventually like to not
rely on magical directories (e.g. the published file may be stored somewhere
else). The client-side query filters on both conditions, so if it's in the
.shared directory OR if it's got the new flag property, it won't show up in the
dashboard list.