Fix longstanding run-enabling-too-early
When we switched to the new standalone style, the call to restartInteractions
that initializes the REPL was moved. This made it so the loader got hidden
before this was called, so the interface encouraged you to run your program
_before_ the initial restartInteractions call happened. Then RI would trigger
and either reset the interaction environment or clash with a
run-while-already-running error.

This fixes it from a UI point of view by making the run button enable and
loader disable at the right time. It's a little gross code-wise because it's a
global query to set these things, but it's at exactly the right point now.