Fix to address https://github.com/brownplt/pyret-lang/issues/1211
It's a bit unfortunate that this is the code that fixes this issue, because
it's hard to understand.

There was a flat-out bug here that was usually masked by the typical times
requests would take, and typical user activity.

The bug happened any time the file picker was used to open an Pyret file while
an un-named, un-saved file was open. A the "create-and-save" chain of promises
(deterministically!) interleaved with the setting of the "current file" to the
newly-opened one, resulting in the contents of the un-named, un-saved file
being written over the newly opened file.

However, a request for the _contents_ of the newly opened file are fetched
before this save triggers. This is _not_ deterministic, and has to do with the
response time from Google. Usually, for whatever reason, the fetch would return
before the save, so the contents would be correctly put into the editor. Then,
if the user did anything that caused a save, they'd never notice an issue
(though there would be an oddball revision with empty contents in the history
of the file). This made the issue pretty hard to detect and reproduce.

If the contents fetch failed or was slow enough to arrive at Google _after_ the
save happened, then the empty contents would be returned and put into the
editor. This was the less frequent race-y behavior that makes the open action
look totally broken.

This commit forces the create-and-save to happen in a chain, so the page's
notion of the current file cannot be updated until the promise is completely
resolved.

There is a separate UX question of whether this save really needs to happen or
should just be a warning. Since I'm pretty confident this fixes the issue, I'm
pushing it and leaving the UX question for more discussion.