Basics of making this work with stack mgmt as library
This isn't yet pretty code; it pretty much litters delimit around until the
important APIs work.

It also skips some amusingly unnecessary code that is trying to print out the
check results of the compiler itself. This seems like a build config issue.

It also completely fails to do the right thing with parsing exception stack
traces (I think). Some of the error rendering infrastructure is off in general
due to copious use of runThunk and pauseStack, so that will need to be handled
later.

Really, all of these delimits should be able to move _inside_ the appropriate
places in Pyret's runtime. For example, it's likely that runThunk should be
responsible for starting a delimit. This would be a useful first refactoring to
do – make it so that CPO doesn't know about delimit, and only knows about
configuring the pause interval and stack depth with $T__.getRTS()

This might make it easier to have the existing render infrastructure work
transparently, since pauseStack/run interleavings will work without
modification.