UI tweaks to improve UX of REPL
The _actual_ performance is improved by https://github.com/brownplt/pyret-lang/commit/68d935563424ce4f48fb211b763533a23e84d8ab

This commit improves the perceived performance (according to me).

The idea is to delay disabling the run button/run indicator for a fixed
period (200ms for now, but this could increase). I think it's
eye-catching that the run button flashes, which leads to a greater
perception that something happens.

Since the run button stays active during this time and the user could
conceivably press Ctrl+Enter or sneak a click in, this commit also adds
state to make sure we don't run when we shouldn't be.

On a Macbook simple interactions are well below 100ms; I'm pushing this
to horizon in part to more easily test this on a Chromebook. The net
effect does feel much more snappy to me.