Handle stars whose step and point counts are not relatively prime.
Polygons and stars share the same underlying constructor, the PolygonImage, so we first need to update PolygonImage to draw multiple disjoint paths.  Next, when constructing stars, we want the points to point upward, so we tweak the adjustment angle to ensure that.  (This is a slight breakage for even-sided stars, but I suspect this is a rare situation...)