Experimental commit, trying to unify the two libraries a bit better.  NOTE: I have not run this code in CPO yet!  (Not quite sure which branch of pyret-lang I need for this, right now.)  All I know is that the individual files at least have no syntax errors; they may not work at all in practice...
make-image.js is a higher-order library that takes in a name (to blame in annotations) and a dictionary of annotation checkers (which can change between the two libraries) and returns a dictionary of Pyret values that construct images.  Then image-typed.js and image-untyped.js are two JS ibraries that instantiate make-image.js with their respective types.

There are several discrepancies between the two libraries, which are a bit easier to see now that the commonalities are factored out into make-image:

* image-typed has a bunch of -list versions of image combinators (overlay, beside, above, etc.) that are essentially variadic versions of their corresponding binary combinators
* image-typed has overlay-onto-offset
* image-typed has move-pinhole
* image-typed has named-color
* image-untyped has color-list-to-bitmap
* image-untyped exports is-x-place, is-y-place, is-image-color and is-mode, which might cause conflicts in re-exporting this at the Pyret level
* The two libraries disagree on the signature for color-list-to-image (should it include a pinhole, or just set it at the center?)
* The two libraries disagree on the signature for name-to-color