Update jwt per https://nvd.nist.gov/vuln/detail/CVE-2016-10555
Based on my reading of the issue, this vulnerability doesn't affect CPO (though
it is a moderately bad bug in the JWT library), because the bug allows an
attacker to forge a JWT token. But we're in this situation since the request is
coming directly from G:

“An ID Token is a JWT (JSON Web Token), that is, a cryptographically signed
Base64-encoded JSON object. Normally, it is critical that you validate an ID
token before you use it, but since you are communicating directly with Google
over an intermediary-free HTTPS channel and using your client secret to
authenticate yourself to Google, you can be confident that the token you
receive really comes from Google and is valid.”

https://developers.google.com/identity/protocols/OpenIDConnect#obtainuserinfo