Add outliers to single-series box plots
Add new calculations to box-plot-from-lists to pass along information about
outliers and "fences", which are the boundaries of the whiskers excluding
outliers.

The G chart library can't handle variable-length rows, so we check if the data
only has a single series. If it does, we add extra outlier columns and render
them as points. If it has multiple rows, we make sure the existing behavior is
maintained.