Urgent fix for a bug @jswrenn and @kfisler reported but has wide impact
Something's up with the GDrive v2 API that's giving a 307 redirect to a bad
CORS endpoint for the downloadUrl we usually use, resulting in failure to load
saved files for logged-in users.

This fix uses the v3 endpoint, which successfully returns the contents of files
in the limited testing I've been able to do. Unfortunately, I don't have a
great docs link to point to here, just some of my historical knowledge of how
these query parameters have worked and matching on what downloadUrl used to
look like (it's not directly provided on the v3 API).