Construct a request directly, skipping GDrive client lib
This tidies up the fix in the previous commit
50fb4ce7328ceedf15ba30bf6044ef4cb085e0b5 that removes the use of webContentLink
from the server, which can cause Google to rate-limit the code.pyret.org (bad
if a whole class logs on to look at the same shared image!). Instead, it makes
an authenticated request directly for the contents.

This was tricky because of some behavior I don't understand in the drive client
library, noted below and in a comment.

If we make the request with `responseType: 'binary'` (or don't specify it, or
use `responseType: 'text'`) for image data, we get back jpeg data containing
junk rather than the start-of-jpeg data (e.g. unicode replacement characters).
I suspect some undesired stringification is happening.

(responseType is not documented that I can find anyway, but it's described in
this issue https://github.com/googleapis/google-api-nodejs-client/issues/1922)

In addition, when I try the Node.js example for downloading files at
https://developers.google.com/drive/api/v3/manage-downloads, I get an error
that `.on` is not a function (neither is `.pipe`), so I'm not sure what's going
on with that API documentation.

So rather than mucking with the client library, just construct the request
ourselves, using the lightly documented but remarkably stable alt=media
parameter (noted in
https://developers.google.com/drive/api/v3/reference/files/get and
https://developers.google.com/drive/api/v3/manage-downloads) and a Bearer
header.