Search replTextOutput and replOutput for selenium repl
TBH, I'm not sure when the distinction was made or if it's been around
for a while and it's just never mattered in a test. But some of the repl
outputs are listed as replOutput and some as replTextOutput, and both
should be considered when scanning selenium results.

This commit also includes a regression test for type-checking
beside-list