Remove some references to pyret-ide
Ultimately I think bitrot got to this and it's not worth it to keep the build
including it. We learned some good lessons for it, and both and and webpack
might still be involved. But I think it's interfering with the build right now,
so remove it since it's unused.