Add some robustness against flaky tests, and identify a potential issue
Add two extra wait steps to these tests to remedy some recent flakiness (e.g.
https://travis-ci.com/github/brownplt/code.pyret.org/builds/206669889)

1. Make sure there is text in the editor before clicking Run. This may be
another symptom of the seemingly hard-to-squash bug of running before programs
are loaded, and may also simply be that Selenium can click a button sooner than
a user. In any event, that could be explored more, but make sure there is
content in the editor before running in the shareUrl tests.

2. Add a 1s sleep before using the REPL. Our previous check of "is the break
button disabled" could quite naturally race -- we don't *enable* the break
button until ~500ms after the user clicks run to avoid a jittery interface on
quick runs. So sleep for a second before trying the interaction. There might be
a more accurate way to do this -- evalPyret already waits for the >>> prompt to
show up again, the challenge is then making sure the program is actually
running before waiting for that to happen. I tried a few things that didn't
seem to work, but the sleep here is (I think) quite reliable. Right after I
push this Travis will likely tell me if I'm wrong.