Seperated bar-chart-series into 2 seperate dataseries
**Current Update**:
Seperated bar-chart-series into
1) **Bar-chart-series**: [bar-chart, freq-bar-chart]
2) **Multi-Bar-Chart-Series**: [Grouped-bar-chart, Stacked-bar-chart]

Also sets up a **colors** method for both dataseries, using a list of colors to
1) color each bar in the Bar-chart-series
2) color each stack or group member in the Multi-bar-chart-series

**Suggestion For Future Development**:
Right now the  **colors** method
1) colors the bar from *Left to Right* in Bar-chart-series and Grouped-bar-charts
2) colors the bar from *Bottom to Top* in Stacked-bar-charts

and sticking to default colors for the rest if unspecified. This doesn't account for the case if users want to color a bar in the middle or end and want the begginning to be the default color, so for version 1.1 that's something we might want to consider.