Default Colors [Marker Colors]
This is setting up the Defaults to the teacher marker colors that Emmanuel recommended.

Personally, I find this color palette to be a little lacking when compared to google's defaults since the [Red, Green, and Blue] are too bright and the shift to Black and Brown seems too sudden.

We can always change/update the color palette later to be a little bit more refined while still prioritizing colors that teachers have in their marker collection.