Add-Pointer Method [Stacked-Bar-Chart] [Unfinished]
Implemented a Tentaive working add-pointer method that takes in a list of tick names, and a list of tick positions and creates "pointers" on a second axis for Stacked Bar Chart.

PROBLEM: When the tick positions are too close together one or both tick labels disappear.

UNFINISHED: I still have to update this to work with grouped bar charts and regular bar charts as well as import the tests from the google doc to the actual testing file. I'll get that done sometime tommorow