Fixed add-pointer, and intervals/error bars + custom axis, and added interval color method
Turns out the new annotations were throwing off the add-pointer method because they were adding a lot of extra role columns so I just set the target of all those columns to the second axis.
I updated the intervals method to work with the custom axis method so that the intervals dont get suddenly cut off from the top and bottom of the screen
Added a interval color method that lets you change the color of the intervals and make it stand out more.