Fix a bug where the editor/repl height were wrong with bonnie menu open.
The minimal change here was to use offsetHeight instead of scrollHeight so we no longer count the height of the absolutely-positioned menu.

A major question I have is in what case this code is needed at all. Seems like the same goal could likely be met in CSS and I don't understand in what case the toolbar is larger than the CSS size and requires running this code dynamically on all scroll/resize events.