Remove unused transitive dependencies of node modules.
Several of the node packages we use have optional dependencies that are currently being bundled, but not used, taking up ~700kb of space in the built version of npm-dependencies.js. This change directs browserify to explicitly not bundle those packages.

Examples:
- colorspaces is a 10kb library for color math, but it is currently pulling in the whole of stylus, a css preprocessor that is 400kb.
- js-md5 and js-sha256 are small js implementations but they try to pull in node's crypto modules to be faster when possible, using ~300kb of code that is never run.

This commit instructs browserify to explicitly never bundle several packages by name. If we ever _do_ want to use those packages, we'll need to undo this change. (But, also, we'll probably have done something wrong.)