Add GOOGLE_SERVER_API_KEY as an environment variable
Since we now use the api key on the server to request public files, we need to
have something with a little more authority than on the client. The client is
restricted to referer headers of the actual domain (like code.pyret.org,
pyret-horizon.herokuapp.com, etc). So allow for two different keys, one that
goes on the page and is restricted, and one that goes on the server and is
unrestricted (and not shared).

Overall this is pretty low-key stuff, the api keys don't have the authority to
do much but read public files without user consent, so this isn't something
that absolutely needs to be kept secret.