Slightly improve sizing.
This version perfectly sizes the Snap! area for the first page load. Resizing
doesn't really work, but see below.

The important bits are:

- Move .replMain to static HTML to make it (and its size) easily accessible
- Set the initial size of the canvas to the .height() and .width() of replMain
- Insert the Snap! editor *inside* replMain

I think the code in the resize observer should work (it runs!). If you remove
the `false` argument to the world Morph constructor, you get some resizing
behavior that might be better (?) but it seems like the bottom is always behind
the footer. It could also be that when you remove the false that's always a bad
idea because it tries to be the size of the page (?) by directly accessing
window. Not sure.

This link has docs at the top and then code for bits that aren't documented.
Poking around in here more will probably have the answer:

https://github.com/jmoenig/morphic.js/blob/master/morphic.js