Choose Context modal doesn't show bad non-context strings.
Previously, if your first line wasn't a use context declaration, the Choose Context modal would just grab an arbitrary substring of your first line, resulting in gibberish.

This change first checks if the first line is a use context. If so, it shows the current context. If not, it shows the current context as blank and lets you enter a new one.