Some guards and logs around initialization and edits
One thing that's challenging is the use of on('change') on CodeMirror,
which triggers for calls to setValue().

In the events API, we change these to all use replaceRange with an
explicit `origin`. However, there are setValue() calls elsewhere in the
codebase. We could ignore these by checking the origin of "setValue",
but this could be a mistake if we do want to propagate around updates
like those. Need to think about this.

For now, filter out ones that come in before we are initialized, and
wait for the reset event to start sending things out. This makes sure
that the initialization of setting e.g. `use context current` doesn't
get sent out as an event.

Sometimes we get "empty" resets from VMT, and this commit just filters
those out. But it's not clear when exactly that happens.