Add config option to proxy shared reqs through another server
This is a total hack, and should probably be disabled. However,
for demo servers it can be incredibly valuable to have share
links from e.g. code.pyret.org work. Currently they won't
because the Google credentials are wrong. But, we can simply
configure these fetches to ask for the corresponding resource on CPO.

Consider trying that, then falling back to looking on this server if needed?