Stop, and wait for results from, the last run
The high-level goal here is a clean solution to having an in-control
user move everyone along if a reactor or chart modal is open.

The idea is to trigger a stop if a user is (a) not in control and (b)
a run or runInteraction event comes in. This means that all users can
interact with the chart, reactor, etc on their own, and won't be
interrupted if the in-control user stops theirs. But as soon as the in-
control user moves on to run another interaction or use the Run button,
it will forcibly stop the other clients.

This is *not* perfect. Other designs might work better.

The best case for this design is that people get told in chat that they
should close on their own, and the forced stop is just a way to sync
back up.

Some weaknesses and tradeoffs discussed below:

1. For a main program like this, if the forced stop happens the `r`
   variable ends up not being defined, which isn't great.

    r = reactor:
      init: 1,
      on-tick: _ + 1
    end

    r.interact()

2. We could capture the “close” event of the modals and try to propagate
   it around. This is tricky because:
   - It's hard to account for all the ways the modal can close (click,
     `stop-when` trigger, reactor ending in error, etc)
   - The in-control user's game might end early (because of stop-when),
     and it could be nice to let others continue.

3. We could trigger something less drastic than a full “stop” event. For
   example, we could just do JSworld.shutdown({ clean: true }) when the
   next interaction happens.

   However, a single program can have multiple reactors that open in
   serial, conditional on what actions happened. This happens in
   practice! For example the car simulator in physics starts several
   reactors for several different initial conditions. And in general
   it's nice to have a solution that for sure gives control back to the
   REPL.

4. Another option would be to:

   - Queue the interactions that come in while the reactor is running,
     and run them once the reactor stops.
   - Only do a force stop when the *Run* button is clicked.

   This would make it so folks can independently run their animations,
   and when they quit the next interactions would start to process. If
   the in control user instead re-runs the whole program, *then* all
   others get stopped. The queue is a bit nontrivial to maintain, but
   this may be a better user experience for when a "re-sync" ought to
   happen.

   The queue is also nice architecturally because there are potential
   issues of things taking different times to run on different machines
   for long-running programs. This could be a potential issue for e.g.
   loading sheets lagging on some folks' computers and causing them
   to get stopped unnecessarily, and the “queue until rerun” would help
   a lot with that.