Queue all effects from messages in a single queue, reset on run
This implements strategy #4 from https://github.com/brownplt/code.pyret.org/commit/af6f425221af5774c3b9d6bda46791f9331db191

The idea is that *all* messages (other than gain/lose control) now go
into a queue. Each message's effect is an async thunk (so it returns a
promise) and there's a standard queue-processing loop for it.

The exception is run events (as in clicking the Run button), which have
the special effect of *stopping the program* and *clearing the queue*.
This makes it so clicking “Run” will bring everyone back in sync.

When a modal is open, nothing will force it closed for clients other
than clicking the run button; interactions typed after it by those
in control will queue up and wait for the active user to finish the
modal evaluation themselves (e.g. closing the window, their stop-when
terminates). The promise used to finish a run or runInteraction event
is tied to the termination of the evaluation, so all the ways we have
of stopping evaluation are supported.

A thought – we could potentially re-enable the stop button for users
who aren't in control for long-running or hung programs with this
scheme, because that can be under independent control.

I think that reset() needs to be rewritten in terms of the queue still,
because its loop over the pending interactions will (I think) continue
to run even after the stop() from clicking run. So if a reset event
is followed by a run event, it won't cleanly clear up all the pending
stuff. This is next up.